Refactor LLM api to separate module
